Update patch set 1
Patch Set 1: Code-Review-2

Just keeping it here at something to look at some time.

We should see if AOSP 10 to AOSP 11 upgrade works fine, and if it does, then I guess we'd have to update the algorithm to deal with additional grids.

We could just leave it alone as well, but with 12 there might be more changes so it's worth just seeing if this works in AOSP or if there's something obvious we missed / if it a quick fix is possible to get this working properly.

Patch-set: 1
Reviewer: Gerrit User 1000000 <1000000@35266860-6252-469c-a67d-185064479fc0>
Label: Code-Review=-2