Update patch set 2
Patch Set 2: Code-Review-1

I'm not sure of the impact of getting rid of the cache completely, especially on a device like the A2.

I don't know if there's a way to check, but if there is, perhaps only do this for work profile apps. (i.e. icons that have a badge)

Patch-set: 2
Reviewer: Gerrit User 1000000 <1000000@35266860-6252-469c-a67d-185064479fc0>
Label: Code-Review=-1